Worked on Synonyms component
Moved the Synonyms component because React was not happy with it being within the Results component, due there being two components being returned which would need different keys. The Synonyms component is now returned within the Meaning component. There is still a bug in that the h4 Synonyms is being returned, even when there are no synonyms in the array.